Refactored auth and pagination
